Add debouncing to form submissions and improve UI. Chakra-ui replaced with Antd
Refactored the currency conversion form in the app to include debouncing (which prevents the form from being submitted too often). This was made possible through lodash. The user interface was also improved by integrating ant-design, which changed the button, select and input elements for enhanced visual representation and performance. The CurrencyRatesSection was temporarily commented out and additional padding was added to the application for improved appearance.